Add ability to initialize user secrets for Directory.Build.props and Directory.Build.targets, introduce fixes according to PR comments
