+++ src/main/kotlin/eu/gillissen/rider/usersecrets/UserSecretsService.kt
@@ -1,5 +1,6 @@
 package eu.gillissen.rider.usersecrets
 
+import com.intellij.openapi.actionSystem.AnActionEvent
 import com.intellij.openapi.project.Project
 import com.intellij.openapi.util.SystemInfo
 import com.intellij.openapi.vfs.VirtualFile
@@ -9,8 +10,8 @@ import java.io.File
 import java.util.concurrent.TimeUnit
 
 object UserSecretsService {
-    val supportedFileExtensions = arrayOf("csproj", "vbproj", "fsproj")
-    val supportedFileNames = arrayOf("Directory.Build.props", "Directory.Build.targets")
+    private val supportedFileExtensions = arrayOf("csproj", "vbproj", "fsproj")
+    private val supportedFileNames = arrayOf("Directory.Build.props", "Directory.Build.targets")
 
     const val UserSecretsIdMsBuildProperty = "UserSecretsId"
 
@@ -22,8 +23,10 @@ object UserSecretsService {
     }
 
     fun initUserSecrets(projectFile: VirtualFile): Int {
-        val projectDirectory = projectFile.toIOFile().absoluteFile.parentFile
-        val output = Runtime.getRuntime().exec("dotnet user-secrets init", null, projectDirectory)
+        val projectIoFile = projectFile.toIOFile()
+        val fullProjectFilePath = projectIoFile.absolutePath
+        val projectDirectory = projectIoFile.absoluteFile.parentFile
+        val output = Runtime.getRuntime().exec("dotnet user-secrets init --project \"${fullProjectFilePath}\"", null, projectDirectory)
         output.waitFor()
 
         return output.exitValue()
@@ -36,9 +39,7 @@ object UserSecretsService {
             .blockingGet(1, TimeUnit.MINUTES)
             ?: return null
 
-        val secretsId = msBuildProperties[UserSecretsService.UserSecretsIdMsBuildProperty] ?: return null
-
-        return secretsId
+        return msBuildProperties[UserSecretsService.UserSecretsIdMsBuildProperty]
     }
 
     fun getUserSecretsDirectoryRoot(): String {
@@ -46,5 +47,26 @@ object UserSecretsService {
             "${System.getenv("APPDATA")}${File.separatorChar}microsoft${File.separatorChar}UserSecrets${File.separatorChar}"
         else
             "${System.getenv("HOME")}${File.separatorChar}.microsoft${File.separatorChar}usersecrets${File.separatorChar}"
+    }
+
+    fun isActionSupported(actionEvent: AnActionEvent): Boolean {
+        val project = actionEvent.getActionProject()
+        if (project == null || project.isDefault) {
+            return false
+        }
+
+        val projectFile = actionEvent.getActionProjectFile()
+        if (projectFile == null || !UserSecretsService.isActionSupportedForFile(projectFile)) {
+            return false
+        }
+
+        return true
+    }
+
+    private fun isActionSupportedForFile(projectFile: VirtualFile?): Boolean {
+        if (projectFile == null) return false
+
+        return UserSecretsService.supportedFileExtensions.any { it.equals(projectFile.extension, ignoreCase = true) } ||
+                UserSecretsService.supportedFileNames.any { it.equals(projectFile.name, ignoreCase = true) }
     }
 }
